Back_end/requirements.txt
Flask
pytest
flake8
gunicorn
psycopg2-binary
Flask-cors